feat: 点位分页压缩回传采取点位范围的分页策略

+++ genshin-map-api/genshin-map-api-core/genshin-map-api-core-core/src/main/java/site/yuanshen/genshin/core/controller/MarkerController.java
@@ -70,7 +70,7 @@ public class MarkerController {
     @Operation(summary = "通过bz2返回点位分页", description = "查询分页点位信息，返回bz2压缩格式的byte数组")
     @GetMapping("/get/list_page_bz2/{index}")
     public byte[] listPageMarkerBy7zip(@RequestHeader(value = "isTestUser", required = false) String isTestUser,
-                                      @PathVariable("index") Long index) throws IOException {
+                                      @PathVariable("index") Integer index) throws IOException {
         return markerDao.listPageMarkerByBz2(StringUtils.hasLength(isTestUser),index);
     }
